+++ Questionable.Model/Questing/EExtraSkipCondition.cs
@@ -26,4 +26,5 @@ public enum EExtraSkipCondition
     /// </summary>
     CostaDelSol,
     NewGamePlus,
+    NotNewGamePlus,
 }