+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - Feature: config window is now resizable -alydev
+- Feature: SwitchClass now supports job>class by removing job stone -alydev
 - Change: Stop command skipped for UI stop and ESC key -alydev
 - Bug fix: interact with crystarium aetheryte in quest even if already unlocked -alydev